+++ Integrations/IntSight/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
 
+## [20.1.2] - 2020-01-22
+  - Improved logging for fetch_incidents.
+  - Improved error handling.
+
 ## [19.12.1] - 2019-12-25
 Fixed an issue with the arguments *is-hidden* and *rate* of the ***intsights-close-alert*** command.