+++ Integrations/MicrosoftTeams/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## [Unreleased]
+
+
+## [20.1.2] - 2020-01-22
+Added the ability to mention users in the ***send-notification*** command.
 Added 2 commands.
   - ***add-user-to-channel***
   - ***create-channel***