+++ Scripts/CommonServerPython/CHANGELOG.md
@@ -1,10 +1,26 @@
 ## [Unreleased]
--
+Added ***remove_empty_elements*** command.
+Added ***datetime_to_string*** command.
+Added ***safe_load_json*** command.
+Added encoding from UTF-8 for Python 2.
+
+## [20.3.1] - 2020-03-04
+Added **DomainGlob** to the **FeedIndicatorType** class.
+Added the *timeline* argument to the ***return_outputs*** convenience function.
+
+## [20.2.3] - 2020-02-18
+Added cveRegex to validate cve_id format.
+
+## [20.2.0] - 2020-02-04
+Added ***ip_to_indicator_type*** command.
+
+
+## [20.1.2] - 2020-01-22
+Added encode string results - safe handle unicode strings to demisto results
 
 ## [20.1.0] - 2020-01-07
  - Added the ***argToBoolean*** command, which takes a input value of type string or boolean and converts it to boolean.
  - Added the **batch** command, which takes an iterable and how many items to return and yields batches of that size.
- 
 
 ## [19.12.0] - 2019-12-10
 -